pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.17",
    "ollama>=0.6.0",
    "python-dotenv>=1.2.1",
]
